+++ include/utime.h
@@ -10,7 +10,7 @@
 /* types */
 # ifndef time_t
 #  define time_t time_t
-typedef int time_t;
+typedef long long time_t;
 # endif
 
 struct utimbuf
